+++ CHANGES.rst
@@ -21,6 +21,8 @@
 - Create a new layer and copy features to it instead of using CopyLayer to work around GDAL 3.9 behaviour (#1094).
 - Display migration warnings in the upgrade log when upgrading a schematisation
 - Added additonal check for legacy geopackages.
+- Collect any errors or warnings from 'Check Schematisation' that are associated to a geometry as vector layer
+- Seperate model checker results into layers by associated geometry
 - Fix setting path in computational grid from schematisation processing algorithm